+++ NEWS.md
@@ -1,6 +1,7 @@
 #wrappedtools 0.9.4
 - ksnormal can handle input of 0 data points
-- new function label_outliers
+- new function label_outliers to extend boxplots
+- new function detect_outliers for compatibility with geom_boxplot
 
 # wrappedtools 0.9.3
 - compare2numvars checks indep. variables for 2 levels